Fixed configuation for verifier build
