+++ org.lflang/META-INF/MANIFEST.MF
@@ -14,10 +14,9 @@ Require-Bundle: org.eclipse.xtext,
  org.eclipse.xtext.xbase.lib;bundle-version="2.27.0",
  org.eclipse.xtext.util,
  org.eclipse.emf.common,
- org.antlr.runtime;bundle-version="[3.2.0,3.2.1)",
+ org.antlr.runtime;bundle-version="4.10.0",
  com.google.guava,
  org.apache.commons.cli;bundle-version="1.4",
- org.lflang.epoch.external.dependencies,
  org.eclipse.lsp4j;bundle-version="0.14.0",
  com.fasterxml.jackson.core.jackson-core,
  com.fasterxml.jackson.core.jackson-annotations,
@@ -33,7 +32,9 @@ Require-Bundle: org.eclipse.xtext,
  org.eclipse.core.resources;bundle-version="3.16.0",
  org.eclipse.lsp4j;bundle-version="0.14.0",
  org.eclipse.lsp4j.jsonrpc;bundle-version="0.14.0",
- org.eclipse.sprotty;bundle-version="0.9.0"
+ org.eclipse.sprotty;bundle-version="0.9.0",
+ org.lflang.epoch.mirror.json;bundle-version="1.20230618.0",
+ org.lflang.epoch.mirror.kotlin;bundle-version="1.6.21"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Export-Package: org.lflang,
  org.lflang.ast,
