+++ org.lflang/build.properties
@@ -2,7 +2,8 @@ source.. = src/,\
            src-gen/,\
            lingua-franca/core/src/main/java/,\
            lingua-franca/core/src/main/kotlin/,\
-           lingua-franca/core/src/main/resources/
+           lingua-franca/core/src/main/resources/,\
+           lingua-franca/core/build/generated-src/antlr/main
 bin.includes = model/generated/,\
                .,\
                META-INF/,\
