+++ org.lflang/pom.xml
@@ -47,8 +47,6 @@
                         <argument>rootPath=/${project.basedir}/..</argument>
                         <argument>-p</argument>
                         <argument>lfSource=${lf-java-sources}</argument>
-                        <argument>-p</argument>
-                        <argument>antlr4Gen=${lf-antlr-generated}</argument>
                     </arguments>
                     <classpathScope>compile</classpathScope>
                     <includePluginDependencies>true</includePluginDependencies>
@@ -89,16 +87,6 @@
                         <groupId>org.eclipse.xtext</groupId>
                         <artifactId>xtext-antlr-generator</artifactId>
                         <version>[2.1.1, 3)</version>
-                    </dependency>
-                    <dependency>
-                        <groupId>org.json</groupId>
-                        <artifactId>json</artifactId>
-                        <version>${jsonVersion}</version>
-                    </dependency>
-                    <dependency>
-                        <groupId>org.antlr</groupId>
-                        <artifactId>antlr4-runtime</artifactId>
-                        <version>${antlr4Version}</version>
                     </dependency>
                 </dependencies>
             </plugin>
