--- org.lflang/src-kotlin-mock/org/lflang/KotlinMock.java
@@ -1,10 +1,0 @@
-package org.lflang;
-
-/**
- * This entire package is for mocking Kotlin
- * classes for a debug run in Eclipse.
- * The actual product will have the real classes.
- */
-public class KotlinMock {
-  public static final String MSG = "Kotlin-based classes are not available in a live run of Epoch, only in the final product.";
-}
