+++ org.lflang/src-mock/org/lflang/generator/cpp/CppFileConfig.java
@@ -5,21 +5,21 @@ import java.nio.file.Path;
 
 import org.eclipse.emf.ecore.resource.Resource;
 import org.lflang.FileConfig;
-import org.lflang.KotlinMock;
+import org.lflang.Mock;
 
 public class CppFileConfig extends FileConfig {
 
     public CppFileConfig(Resource resource, Path srcGenBasePath,
             boolean useHierarchicalBin) throws IOException {
         super(resource, srcGenBasePath, useHierarchicalBin);
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
     public Path getIncludePath() {
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
     public String getRuntimeIncludePath() {
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 }
