+++ org.lflang/src-mock/org/lflang/generator/ts/TSFileConfig.java
@@ -1,25 +1,25 @@
-package org.lflang.generator.rust;
+package org.lflang.generator.ts;
 
 import java.io.IOException;
 import java.nio.file.Path;
 
 import org.eclipse.emf.ecore.resource.Resource;
 import org.lflang.FileConfig;
-import org.lflang.KotlinMock;
+import org.lflang.Mock;
 
-public class RustFileConfig extends FileConfig {
+public class TSFileConfig extends FileConfig {
 
-    public RustFileConfig(Resource resource, Path srcGenBasePath,
+    public TSFileConfig(Resource resource, Path srcGenBasePath,
             boolean useHierarchicalBin) throws IOException {
         super(resource, srcGenBasePath, useHierarchicalBin);
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
     public Path getIncludePath() {
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
     public String getRuntimeIncludePath() {
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 }
