+++ org.lflang/src-mock/org/lflang/generator/ts/TSGenerator.java
@@ -1,29 +1,29 @@
 
-package org.lflang.generator.rust;
+package org.lflang.generator.ts;
 
-import org.lflang.KotlinMock;
+import org.lflang.Mock;
 import org.lflang.Target;
 import org.lflang.generator.GeneratorBase;
 import org.lflang.generator.LFGeneratorContext;
 import org.lflang.generator.TargetTypes;
 import org.lflang.scoping.LFGlobalScopeProvider;
 
-public class RustGenerator extends GeneratorBase {
+public class TSGenerator extends GeneratorBase {
 
-    public RustGenerator(LFGeneratorContext context,
+    public TSGenerator(LFGeneratorContext context,
             LFGlobalScopeProvider scopeProvider) {
         super(context);
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
     @Override
     public TargetTypes getTargetTypes() {
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
     @Override
     public Target getTarget() {
-        throw new IllegalStateException(KotlinMock.MSG);
+        throw new IllegalStateException(Mock.KOTLIN_MSG);
     }
 
 }
