+++ pom.xml
@@ -18,13 +18,16 @@
         <!-- Sync with build.gradle -->
         <commonsCliVersion>1.4</commonsCliVersion>
         <!-- Note that this version is tied to our version of Gradle -->
-        <kotlinVersion>1.6.20</kotlinVersion>
+        <kotlinVersion>1.6.21</kotlinVersion>
 
         <!-- Eclipse specific components -->
         <!-- Important: Any version specified for these plugins must be available via the update sites listed in the target platform definition. If not, the Epoch build will fail! You can find the target platform definition in org.lflang.targetplatform/org.lflang.targetplatform.target. Changes to the target platform should also be transferred to the oomph setup (oomph/LinguaFranca.setup). -->
         <xtextVersion>2.31.0</xtextVersion>
         <lsp4jVersion>0.21.0</lsp4jVersion> <!-- This plugin is provided by Xtext -->
         <mwe2LaunchVersion>2.14.0</mwe2LaunchVersion> <!-- This plugin is provided by Eclipse, not Xtext -->
+
+        <!-- ANTLR4 version used in the LF Verifier (not the one used in xtext) -->
+        <antlr4Version>4.7.2</antlr4Version>
 
         <!-- Tycho version used by Maven (Eclipse specific build support)-->
         <tychoVersion>2.7.5</tychoVersion>
@@ -38,12 +41,6 @@
         <!-- Non-version properties -->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <tycho.disableP2Mirrors>true</tycho.disableP2Mirrors>
-
-        <!-- JSON version used in the LF Verifier -->
-        <jsonVersion>20230618</jsonVersion>
-
-        <!-- ANTLR4 version used in the LF Verifier (not the one used in xtext) -->
-        <antlr4Version>4.7.2</antlr4Version>
     </properties>
 
     <profiles>
